+++ src/main/java/com/atwoz/member/domain/oauth/OAuthRequester.java
@@ -1,7 +1,7 @@
-package com.atwoz.member.application.auth;
+package com.atwoz.member.domain.oauth;
 
 import com.atwoz.member.infrastructure.oauth.dto.MemberInfo;
-import com.atwoz.member.ui.auth.support.oauth.OAuthProperties.OAuthProvider;
+import com.atwoz.member.ui.auth.support.oauth.OAuthProvider;
 
 public interface OAuthRequester {
