+++ src/main/java/com/atwoz/member/ui/auth/AuthController.java
@@ -4,7 +4,7 @@ import com.atwoz.member.application.auth.AuthService;
 import com.atwoz.member.application.auth.dto.LoginRequest;
 import com.atwoz.member.ui.auth.dto.TokenResponse;
 import com.atwoz.member.ui.auth.support.oauth.OAuthAuthority;
-import com.atwoz.member.ui.auth.support.oauth.OAuthProperties.OAuthProvider;
+import com.atwoz.member.ui.auth.support.oauth.OAuthProvider;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
